+++ src/blockly/components/LeaveMaze.js
@@ -12,7 +12,6 @@ Blockly.Blocks[LEAVE_MAZE] = {
     this.appendDummyInput()
         .setAlign(Blockly.ALIGN_CENTRE)
         .appendField(Blockly.Msg.LEAVE_MAZE_TITLE);
-    this.contextMenu = false;
     this.setColour(colors.ACTION);
     this.setNextStatement(true, null);
     this.setPreviousStatement(true, null);